requirements.txt
dnspython==2.7.0
numpy==2.0.2
pandas==2.3.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
